[AscendNPU-IR][NormalizeVector] Use the inverse permutation map to compute gather indices
permutation map maps the memref index space onto the result vector index
space, so composing the result coordinates with it yields wrong memref
indices for 3-D case, e.g. (d0, d1, d2) -> (d1, d2, d0).

+++ bishengir/lib/Dialect/HFusion/Transforms/RemoveRedundantWriteAndReadPair.cpp
@@ -157,7 +157,7 @@ struct FoldWidenedTransferReadAfterWrite
       activeSizes = maskSizes;
     }
 
-    for (size_t dim = 0, rank = readType.getRank(); dim < rank; ++dim) {
+    for (int64_t dim = 0, rank = readType.getRank(); dim < rank; ++dim) {
       if (activeSizes[dim] > writeType.getDimSize(dim) ||
           writeType.getDimSize(dim) > readType.getDimSize(dim))
         return failure();
